JAMES-2665 MailRepositoryConfiguration should expose a default protocol
This is needed because the default protocol to be used depends directly of
the configuration provided by the admin, and vary depending of products.

Hence  we need to dynamically retrieve it.

Note that we can introduce a specific configuration field later in mailrepositorystore.xml
later on to avoid implicit knowledge of how to configure the default protocol.