JAMES-2806 ObjectStorage Swift retry one more time when fails to save
Swift cause a big time wait for the first put blob request to response.

Assuming there are 2 thread trying to save a blob in the same time,
and both are expected to fail and when a thread fall back to retry and
create the bucket, the other thread comes to the step of checking
bucket existing, now the bucket is already created then retry mechanism
is rejected.

I want to have the last retry regardless bucket existence to increase
the chance to save a blob. This is just a conner case and rarely we never
come up with retrying saving blob. Except the only one use case of Vault
switching new Bucket to save deleted message.

Then I have to take care about this issue rather than relax
test concurrency level or disable it.