JAMES-2838 CassandraBlobStore deleteBucket implementation
Due to the primary key schema we can not get in a single request all of the blobs part of a bucket.

Thus, we need to scroll all blobs in order to filter out those not part of the deleted bucket.

Note that large rows issues disallow the trivial (PK: bucketName, CK: blobId) design.

A complex timeserie could be imagined in order to get all blobs of a bucket, but iteration over the bucket will still be needed (part of the primary key). Thus, deleting a bucket in the CassandraBlobStore is a complex operation.