JAMES-2704 Random storing can leave some mailboxes empty
This mailet assigns emails between 4 to 8 mailboxes. However, there is no
coordination, hence some mailboxes can be left empty.

This do not plays well with the cache on routing info in that mailet: some
auto provisionned inboxes or not yet created mailboxes might get ignored.

Furthermore, that assertion was hidden 3 method call deep the callstack, violating the rule
"one assertion per test, explicit and at the end on the main testing method".

As such it deserves to be removed.

Note: the main assertion "given 100 mail, we will assign and store it between
400 and 800 times" still stands.