Add a build argument for Spring Dockerfile version
Upon a release we need to change simultaneously maven version and the one
within Spring dockerfile (which relies on hard-coded version positionned
by the assembly plugin). This results in broken build post-release as well
as difficulties into building a specific tag when forgotten.

With this changeset we can specify a build argument upon Spring docker
image creation from CI environment, in order to use the right version
without needing code changes. As version can be detected from build files,
this means that we will avoid that pitfall.