PROTOCOLS-36 [Refactoring] AbstractProtocolTransport response queue is no longer needed
As stated by PROTOCOLS-36:

```
If we mix FutureResponse and Response implementation we MUST ensure that
the right order of responses is maintained. This basicly means that we need
to put responses in a queue and dequeue them in the right order (respecting
the FutureResponse async nature)
```

Removing FutureResponse solves this, and removes the need for a queue.