[Refactoring] Get rid of ImapSessionUtils
This class was used to help retrieving MailboxSession and Username which
is a very common operation.

Having a separeted utility makes the API not straightforward to work with.

I propose leveraging default methods to enhance the developer experience
and allow an ImapSession::getMailboxSession and a ImapSession::getUsernames
methods.

Note that fixing mocked tests was non trivial hence I replaced mocking by
usage of FakeMailboxSession.